Intervals class template
With implementations exported for Float, Int, Time.  This is well
suited for carrying single bit "cuts" or "flags" information.  Imagine
defining a series of time intervals (IntervalsTime) that are of
interest (or are to be avoided), and then converting those into sample
ranges (IntervalsInt) to use on the data vectors.