quicksim.py module (python)
This is a starting point for generating simple G3Frame stream TOD sims
using a G3Pipeline and the appropriate modules.  In the example at the
bottom, the pipeline is seeded with a focal plane and observation
description (a constant elevation scan), and then goes on to generate
full pointing and detector data structures (albeit without any
interesting signal injected).