Library reports an auto-computed version string
This is based on versioneer code, but adapted for CMake builds.  A new
target, so3g-version, is introduced to create a header file defining
the version string.  This target is a dependency of the main library
build.  The version string is computed based on proximity to tags of
the form v* (e.g. v0.0.1).  If the code is exported, the computed
version will be invalid unless the exported version bears a v* tag.
But any builds done in the git tree will have a more informative
version, tag.