Pointer<CoordQuatCyl> and Pointer<CoordQuatZen>
This is a work in progress, but the basic math is in place.  There are
issues of sign convention and polarization direction to resolve.  It's
also disappointingly slower than the flat case.  The test_pointing.py
and test_pointing2.py scripts test all three Pointers.

Also ProjectionEngine::coords and ::pixels can create their own output
arrays.