Uses lists of arrays instead of a single array...
... for signal, pixels, coords.  In some systems, time-domain arrays
are stored in lists or hashes so we can't insist on a single large
array.  Note that in this implementation the user can pass in just
about anything to be used for output, as long as it has n_det items,
and each item exposes a buffer of the right shape.