proj: optional per-detector weights array in Forward
(And also ForwardWeight.)  Updates the demo to make use of this
feature.  A per-detector weight capability makes it easy to handle the
noise model where the noise spectrum in each detector is independent
and white but possibly a different level from other detectors.  It
also allows one to bin up weight maps, if one can define a scalar
weight for each detector sample.