proj: CelestialSightLine.for_lonlat; Assembly can take G3VectorQuat
New constructor helps to create transfers from focal plane coordinates
directly to some celestial pointing; that's useful for sims where you
don't want to go through horizon coordinates.

For Assembly, accept a G3VectorQuat in place of Assembly.  I hesitate
to say that CelestialSightLine is on its way out, because it might
still have extensions for multi-boresight pointing.  But this is
helpful when coordinate hacking in sims and simple cases.