hk: fix .simple() behavior when an empty vector is present
In cases where a time range did not have any data for a requested
field, .get_data simply does not include that result, and .simple()
was then failing to construct the return list, which was supposed to
have the same length as the list of requested fields.

Also clarify behavior of a couple functions when data is not found.