C++: fix how "optional" array arguments are decoded
Prior to this fix, if user passed an argument that was not a numpy
array, the code would treat the argument as having not been passed.
Now, passing an invalid argument (such as a list) will cause an
exception to be raised.  To not pass an optional thing, pass it as
None.