Fix spt3g build to never link to python.
Compiled python extensions distributed in wheels should never
link directly to the core Python libraries, since that would
attempt to bundle the python libs on the build system into the
wheel, rather than resolving the linking at run time after the
wheel is installed.  We already patch the spt3g cmake files to
not search for python development libraries.  This commit
updates the patch to also remove building of compiled unit
tests and examples, which would require linking to libpython.