Fix segfault in recent wheel builds
The 0.2.1 tag introduced ceres-solver + Eigen + GLOG into the
compiled dependencies and into the wheel building scripts.  The
resulting wheels, when installed into a conda environment,
triggered a segfault inside the conda-forge glog package.  I am
not sure the cause of this, since the glog library vendored in
the so3g wheels should have its name modified to avoid such
conflicts.

This work changes the compilation inside the wheels to build
and link static versions of glog and ceres-solver.  I tested
a locally-built (which cibuildwheel) wheel and verified that
the segfault is now gone and that sotodlib master branch unit
tests pass using this so3g wheel.